pkg: extract objectupdate from objectstate
Initially, the `objectstate` package was meant to hold the current
and desired state of the objects the operator manages.
Out of practicality (and a bit of laziness) we then added function to
update some of the objects. It's now time to move these updating
functions in their own package, with no intended changes in behaviour
bar trivial import/renames.
