+++ pkg/numaresourcesscheduler/objectstate/sched/sched.go
@@ -18,18 +18,15 @@ package sched
 
 import (
 	"context"
-	"fmt"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
-	"k8s.io/klog/v2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/k8stopologyawareschedwg/deployer/pkg/manifests"
 
 	nrsv1alpha1 "github.com/openshift-kni/numaresources-operator/api/numaresourcesoperator/v1alpha1"
-	"github.com/openshift-kni/numaresources-operator/pkg/hash"
 	schedmanifests "github.com/openshift-kni/numaresources-operator/pkg/numaresourcesscheduler/manifests/sched"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate"
 	"github.com/openshift-kni/numaresources-operator/pkg/objectstate/compare"
@@ -135,22 +132,6 @@ func FromClient(ctx context.Context, cli client.Client, mf schedmanifests.Manife
 	return ret
 }
 
-func UpdateDeploymentImageSettings(dp *appsv1.Deployment, userImageSpec string) {
-	// There is only a single container
-	cnt := &dp.Spec.Template.Spec.Containers[0]
-	cnt.Image = userImageSpec
-	klog.V(3).InfoS("Exporter image", "reason", "user-provided", "pullSpec", userImageSpec)
-}
-
-func UpdateDeploymentConfigMapSettings(dp *appsv1.Deployment, cmName, cmHash string) {
-	template := &dp.Spec.Template // shortcut
-	template.Spec.Volumes[0] = newSchedConfigVolume(SchedulerConfigMapVolumeName, cmName)
-	if template.Annotations == nil {
-		template.Annotations = map[string]string{}
-	}
-	template.Annotations[hash.ConfigMapAnnotation] = cmHash
-}
-
 func DeploymentNamespacedNameFromObject(obj client.Object) (nrsv1alpha1.NamespacedName, bool) {
 	res := nrsv1alpha1.NamespacedName{
 		Namespace: obj.GetNamespace(),
@@ -186,46 +167,7 @@ func SchedulerNameFromObject(obj client.Object) (string, bool) {
 	return "", false
 }
 
-func UpdateSchedulerName(cm *corev1.ConfigMap, name string) error {
-	if name == "" {
-		return fmt.Errorf("not allow to set an empty name for scheduler in ConfigMap: %s/%s", cm.Namespace, cm.Name)
-	}
-
-	if cm.Data == nil {
-		return fmt.Errorf("no data found in ConfigMap: %s/%s", cm.Namespace, cm.Name)
-	}
-
-	data, ok := cm.Data[SchedulerConfigFileName]
-	if !ok {
-		return fmt.Errorf("no data key named: %s found in ConfigMap: %s/%s", SchedulerConfigFileName, cm.Namespace, cm.Name)
-	}
-
-	schedCfg, err := manifests.KubeSchedulerConfigurationFromData([]byte(data))
-	if err != nil {
-		return err
-	}
-
-	for i, schedProf := range schedCfg.Profiles {
-		// if we have a configuration for the NodeResourceTopologyMatch
-		// this is a valid profile
-		for _, plugin := range schedProf.PluginConfig {
-			if plugin.Name == SchedulerPluginName {
-				schedCfg.Profiles[i].SchedulerName = &name
-			}
-		}
-	}
-
-	byteData, err := manifests.KubeSchedulerConfigurationToData(schedCfg)
-	if err != nil {
-		return err
-	}
-
-	cm.Data[SchedulerConfigFileName] = string(byteData)
-	return nil
-
-}
-
-func newSchedConfigVolume(schedVolumeConfigName, configMapName string) corev1.Volume {
+func NewSchedConfigVolume(schedVolumeConfigName, configMapName string) corev1.Volume {
 	return corev1.Volume{
 		Name: schedVolumeConfigName,
 		VolumeSource: corev1.VolumeSource{
